Consolidate mob builder into npc_builder
